requirements.txt
vk-api==11.10.0
python-dotenv==1.1.1
python-telegram-bot==13.15